requirements.txt
flask==2.0.1
flask-login==0.5.0
flask-sqlalchemy==2.5.1
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.47.0
beautifulsoup4==4.9.3
python-dotenv==0.19.0
gunicorn==20.1.0
psycopg2-binary==2.9.3 